refactor: Move the formula configuration in the actor module
